+++ server/db/schema/profiles.ts
@@ -7,3 +7,4 @@ export const Profiles = pgTable('profiles', {
   createdAt: timestamp('created_at').defaultNow().notNull(),
   updatedAt: timestamp('updated_at').defaultNow().notNull(),
 })
+